+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## Crypto Version [1.5.3]
+
+Aug 18, 2021
+
+### Dependencies
+
+- Crypto 1.5.3.
+- Key 1.5.3.
+- User 1.5.3.
+
+### Behavior Changes
+
+- Set X25519 as default Key type for generation (removed RSA).
+- Changed Key identity as: email <email>.
+
 ## Presentation Version [1.5.4]
 
 Aug 20, 2021